requirements.txt
aiohttp==3.8.1
discord.py==2.1.0
matplotlib==3.6.2
numpy==1.23.5
pandas==1.5.2

google-cloud-storage==2.6.0
gspread==5.7.1
oauth2client==4.1.3

tzdata==2022.7